Increased the size of encoded packet size to remove a compilation warning and make the code more coherent. The interface was wrong, it should be the destructor defined, not the constructor.

+++ src/gui/statisticswindow.cpp
@@ -195,7 +195,7 @@ void StatisticsWindow::presentPackage(uint32_t peer, QString type)
   }
 }
 
-void StatisticsWindow::addEncodedPacket(QString type, uint16_t size)
+void StatisticsWindow::addEncodedPacket(QString type, uint32_t size)
 {
   if(type == "video" || type == "Video")
   {
